requirements-addon.txt
paho-mqtt
pyyaml
opcua
requests
